Slider.js
Slider project using Javascript.
I'm not sure of the good maintainability of my method. I decided not to create my cards in html but rather to replace the properties of the existing one.
It would involve modifying the app.js and adding a condition if you want to add a new card...
CSS possible improvments are mainly on animation but I don't want to spend too much energy on this atm.